I added a script to clean up tmp files. Eventually I will write to use find & clean-up all tmp files & directories.  Right now it only removes twitux cached avatars.  I also updated my git commit helper script so that it will remove '.*.swp' files. These are generally created by vim, GTK-PHP-IDE, gedit or etc when editing a file.  The swp file is left in place it just removed from the current commit.  I have made the removale of swp files by my vim session script optional.  It now takes one option: '--clean-up' if this is set than swp files are removed, if not a session file is still created but the swp files left in place.

+++ helper-scripts/vim:session:make.tcsh
@@ -1,5 +1,12 @@
 #!/bin/tcsh -f
 set session_exec = "vim.session.tcsh"
+
+set clean_up = "no"
+if ( "${?1}" == "1" && "${1}" == "--clean-up" ) then
+	set clean_up = "yes"
+	shift
+endif
+
 set search_dir = "./"
 if ( "${?1}" != "0" && "${1}" != "" && -d "${1}" ) set search_dir = "${1}"
 
@@ -9,11 +16,9 @@ foreach swp ( "`find ${search_dir} -iregex .\*\.swp`" )
 		set session_started = "[done]"
 		printf '#\!/bin/tcsh -f\nvim-enhanced -p ' >! "${session_exec}"
 	endif
+
 	printf "%s" "${swp}" | sed 's/\(.*\)\/\.\(.*\.opml\).swp/"\1\/\2"\ /g' >> "${session_exec}"
-	rm "${swp}"
+	if ( "${clean_up}" == "--clean-up" ) rm "${swp}"
 end
 
-if ( "${session_started}" == "[done]" ) then
-	printf "\n" >> "${session_exec}"
-	chmod +x "${session_exec}"
-endif
+if ( "${session_started}" == "[done]" ) chmod +x "${session_exec}"
